+++ src/main.tsx
@@ -10,7 +10,7 @@ import { PageTitleProvider } from "lib/page-title";
 import ErrorPage from "pages/error/error-page";
 import { StrictMode, Suspense } from "react";
 import { createRoot } from "react-dom/client";
-import Router from "router";
+import Router from "router/routes";
 import "style/_main.scss";
 
 const rootElement = document.getElementById("root");
